refactor: - rearrange data transfer objects and entities in packages - split actor repository from details repository
