--- data/src/main/java/com/london/data/EmptyClass.kt
@@ -1,7 +1,0 @@
-package com.london.data
-
-/**
- * Delete this once a new file is added to this package.
- */
-
-class EmptyClass
