+++ data/src/main/java/com/london/data/datasource/remote/details/actordetails/model/actorimage/ActorImageResponse.kt
@@ -1,0 +1,10 @@
+package com.london.data.datasource.remote.details.actordetails.model.actorimage
+
+import kotlinx.serialization.SerialName
+
+data class ActorImageResponse(
+    @SerialName("id")
+    val id: Int,
+    @SerialName("profiles")
+    val profiles: List<ProfileDto>
+)
