+++ data/src/main/java/com/london/data/datasource/remote/details/actordetails/model/actorimage/ProfileDto.kt
@@ -1,13 +1,6 @@
-package com.london.data.datasource.remote.details.actordetails.model
+package com.london.data.datasource.remote.details.actordetails.model.actorimage
 
 import kotlinx.serialization.SerialName
-
-data class ActorImageResponse(
-    @SerialName("id")
-    val id: Int,
-    @SerialName("profiles")
-    val profiles: List<ProfileDto>
-)
 
 data class ProfileDto(
     @SerialName("aspect_ratio")
@@ -25,4 +18,3 @@ data class ProfileDto(
     @SerialName("width")
     val width: Int
 )
-
