+++ data/src/main/java/com/london/data/mapper/actordetails/ActorTvShowDetailsMapper.kt
@@ -1,11 +1,11 @@
 package com.london.data.mapper.actordetails
 
-import com.london.data.datasource.remote.details.actordetails.model.ActorTvShowCastMember
-import com.london.data.datasource.remote.details.actordetails.model.ActorTvShowCrewMember
-import com.london.data.datasource.remote.details.actordetails.model.ActorTvShowDetailsResponse
-import com.london.domain.entity.actordetails.ActorTvShowCastMemberEntity
-import com.london.domain.entity.actordetails.ActorTvShowCrewMemberEntity
-import com.london.domain.entity.actordetails.ActorTvShowDetails
+import com.london.data.datasource.remote.details.actordetails.model.actortvshowdetails.ActorTvShowCastMember
+import com.london.data.datasource.remote.details.actordetails.model.actortvshowdetails.ActorTvShowCrewMember
+import com.london.data.datasource.remote.details.actordetails.model.actortvshowdetails.ActorTvShowDetailsResponse
+import com.london.domain.entity.actordetails.actortvshow.ActorTvShowCastMemberEntity
+import com.london.domain.entity.actordetails.actortvshow.ActorTvShowCrewMemberEntity
+import com.london.domain.entity.actordetails.actortvshow.ActorTvShowDetails
 
 fun ActorTvShowDetailsResponse.toEntity(): ActorTvShowDetails {
     return ActorTvShowDetails(
