+++ domain/src/main/java/com/london/domain/entity/actordetails/actorimage/ActorImageDetails.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.actordetails.actorimage
+
+
+data class ActorImageDetails(
+    val id: Int,
+    val profiles: List<ImageDetails>
+)
