+++ domain/src/main/java/com/london/domain/entity/actordetails/actormovie/ActorMovieDetails.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.actordetails.actormovie
+
+data class ActorMovieDetails(
+    val id: Int,
+    val cast: List<ActorMovieCastMemberEntity>,
+    val crew: List<ActorMovieCrewMemberEntity>
+)
