+++ domain/src/main/java/com/london/domain/entity/actordetails/actortvshow/ActorTvShowDetails.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.actordetails.actortvshow
+
+data class ActorTvShowDetails(
+    val id: Int,
+    val cast: List<ActorTvShowCastMemberEntity>,
+    val crew: List<ActorTvShowCrewMemberEntity>
+)
