+++ domain/src/main/java/com/london/domain/usecase/actordetailsusecase/GetActorDetailsById.kt
@@ -1,9 +1,9 @@
 package com.london.domain.usecase.actordetailsusecase
 
-import com.london.domain.repository.DetailsRepository
+import com.london.domain.repository.ActorRepository
 
 class GetActorDetailsById(
-    private val repository: DetailsRepository
+    private val repository: ActorRepository
 ) {
     suspend fun invoke(actorId: Int) = repository.getActorDetailsById(actorId)
 }
